fixes and small changes
Fixed login, seperated out the movieroutes, and changed the vote average from a 10 point scale to a 5 point scale
